integer overflow fixed

+++ C++/a.cpp
@@ -1,13 +1,13 @@
 #include <bits/stdc++.h>
 using namespace std;
-const int MIN = 1e9;
+const long MIN = 1e9;
 
 int main() {
     int t; cin>>t;
     while(t--){
         int n; cin>>n;
-        int a[n], b[n];
-        int am = MIN, bm = MIN;
+        long a[n], b[n];
+        long am = MIN, bm = MIN;
         for(int i=0; i<n; i++){
             cin>>a[i];
             if(am > a[i])
@@ -18,9 +18,9 @@ int main() {
             if(bm > b[i])
 				bm = b[i];
         }
-        int sum = 0;
+        long long sum = 0;
         for(int i=0; i<n; i++){
-            sum = sum + (a[i]-am) + (b[i]-bm);
+            sum = (long long)sum + (a[i]-am) + (b[i]-bm);
         }
         cout<<sum<<endl;
